Moved agent status to separate widget
