+++ server/webapp/WEB-INF/rails/webpack/views/pages/new-agents/spec/agent_status_widget_spec.tsx
@@ -1,0 +1,95 @@
+/*
+ * Copyright 2019 ThoughtWorks, Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+import {Agent} from "models/new-agent/agents";
+import {AgentsTestData} from "models/new-agent/spec/agents_test_data";
+import {AgentStatusWidget} from "views/pages/new-agents/agent_status_widget";
+import {TestHelper} from "views/pages/spec/test_helper";
+import m from "mithril";
+import Stream from "mithril/stream";
+
+describe("AgentStatusWidget", () => {
+  const helper = new TestHelper();
+  let buildDetailsForAgent: Stream<string>;
+  beforeEach(() => {
+    buildDetailsForAgent = Stream("");
+  });
+  afterEach(helper.unmount.bind(helper));
+
+  it("should render agent status", () => {
+    const agent = Agent.fromJSON(AgentsTestData.disabledBuildingAgent());
+
+    mount(agent, buildDetailsForAgent);
+
+    expect(helper.byTestId(`agent-status-of-${agent.uuid}`)).toContainText("Disabled (Building)");
+  });
+
+  it("should show a hyperlink when agent is building", () => {
+    const agent = Agent.fromJSON(AgentsTestData.buildingAgent());
+    mount(agent, buildDetailsForAgent);
+
+    const statusCell = helper.byTestId(`agent-status-of-${agent.uuid}`);
+    expect(helper.q("a", statusCell)).toContainText("Building");
+
+    const buildDetails     = helper.byTestId(`agent-build-details-of-${agent.uuid}`);
+    const buildDetailsItem = helper.qa("a", buildDetails);
+    expect(buildDetailsItem).toHaveLength(3);
+    expect(buildDetailsItem[0]).toContainText("up42");
+    expect(buildDetailsItem[0]).toHaveAttr("href", "pipeline_url");
+    expect(buildDetailsItem[1]).toContainText("up42_stage");
+    expect(buildDetailsItem[1]).toHaveAttr("href", "stage_url");
+    expect(buildDetailsItem[2]).toContainText("up42_job");
+    expect(buildDetailsItem[2]).toHaveAttr("href", "job_url");
+  });
+
+  it("should show build details info on click of status", () => {
+    const agent = Agent.fromJSON(AgentsTestData.buildingAgent());
+    mount(agent, buildDetailsForAgent);
+    expect(helper.byTestId(`agent-build-details-of-${agent.uuid}`)).not.toBeVisible();
+
+    helper.click(helper.byTestId(`agent-status-text-${agent.uuid}`));
+
+    expect(helper.byTestId(`agent-build-details-of-${agent.uuid}`)).toBeVisible();
+  });
+
+  it("should hide the build details when clicked on status twice", () => {
+    const agent = Agent.fromJSON(AgentsTestData.buildingAgent());
+    mount(agent, buildDetailsForAgent);
+    expect(helper.byTestId(`agent-build-details-of-${agent.uuid}`)).not.toBeVisible();
+
+    helper.click(helper.byTestId(`agent-status-text-${agent.uuid}`));
+    expect(helper.byTestId(`agent-build-details-of-${agent.uuid}`)).toBeVisible();
+
+    helper.click(helper.byTestId(`agent-status-text-${agent.uuid}`));
+    expect(helper.byTestId(`agent-build-details-of-${agent.uuid}`)).not.toBeVisible();
+  });
+
+  // it("should hide the build details when clicked outside", () => {
+  //   const agent  = Agent.fromJSON(AgentsTestData.buildingAgent());
+  //   const agents = new Agents([agent]);
+  //   mount(agents);
+  //   helper.click(helper.byTestId(`agent-status-text-${agent.uuid}`));
+  //   expect(helper.byTestId(`agent-build-details-of-${agent.uuid}`)).toBeVisible();
+  //
+  //   helper.click(helper.byTestId(`agent-hostname-of-${agent.uuid}`));
+  //
+  //   expect(helper.byTestId(`agent-build-details-of-${agent.uuid}`)).not.toBeVisible();
+  // });
+
+  function mount(agent: Agent, buildDetailsForAgent: Stream<string>) {
+    helper.mount(() => <AgentStatusWidget agent={agent} buildDetailsForAgent={buildDetailsForAgent}/>);
+  }
+});
