+++ server/webapp/WEB-INF/rails/webpack/views/pages/new-agents/spec/agents_widget_spec.tsx
@@ -150,77 +150,6 @@ describe("NewAgentsWidget", () => {
     });
   });
 
-  describe("AgentStatus", () => {
-    it("should render agent status", () => {
-      const agent  = Agent.fromJSON(AgentsTestData.disabledBuildingAgent());
-      const agents = new Agents([agent]);
-
-      mount(agents);
-
-      expect(helper.byTestId(`agent-status-of-${agent.uuid}`)).toContainText("Disabled (Building)");
-    });
-
-    it("should show a hyperlink when agent is building", () => {
-      const agent  = Agent.fromJSON(AgentsTestData.buildingAgent());
-      const agents = new Agents([agent]);
-
-      mount(agents);
-
-      const statusCell = helper.byTestId(`agent-status-of-${agent.uuid}`);
-      expect(helper.q("a", statusCell)).toContainText("Building");
-
-      const buildDetails     = helper.byTestId(`agent-build-details-of-${agent.uuid}`);
-      const buildDetailsItem = helper.qa("a", buildDetails);
-      expect(buildDetailsItem).toHaveLength(3);
-      expect(buildDetailsItem[0]).toContainText("up42");
-      expect(buildDetailsItem[0]).toHaveAttr("href", "pipeline_url");
-      expect(buildDetailsItem[1]).toContainText("up42_stage");
-      expect(buildDetailsItem[1]).toHaveAttr("href", "stage_url");
-      expect(buildDetailsItem[2]).toContainText("up42_job");
-      expect(buildDetailsItem[2]).toHaveAttr("href", "job_url");
-    });
-
-    it("should show build details info on click of status", () => {
-      const agent  = Agent.fromJSON(AgentsTestData.buildingAgent());
-      const agents = new Agents([agent]);
-      mount(agents);
-      expect(helper.byTestId(`agent-build-details-of-${agent.uuid}`)).not.toBeVisible();
-
-      helper.click(helper.byTestId(`agent-status-text-${agent.uuid}`));
-
-      expect(helper.byTestId(`agent-build-details-of-${agent.uuid}`)).toBeVisible();
-    });
-
-    it("should hide the build details when clicked on status twice", () => {
-      const agent  = Agent.fromJSON(AgentsTestData.buildingAgent());
-      const agents = new Agents([agent]);
-      mount(agents);
-      expect(helper.byTestId(`agent-build-details-of-${agent.uuid}`)).not.toBeVisible();
-
-      helper.click(helper.byTestId(`agent-status-text-${agent.uuid}`));
-
-      expect(helper.byTestId(`agent-build-details-of-${agent.uuid}`)).toBeVisible();
-
-      helper.click(helper.byTestId(`agent-status-text-${agent.uuid}`));
-
-      expect(helper.byTestId(`agent-build-details-of-${agent.uuid}`)).not.toBeVisible();
-    });
-
-    // it("should hide the build details when clicked outside", () => {
-    //   const agent  = Agent.fromJSON(AgentsTestData.buildingAgent());
-    //   const agents = new Agents([agent]);
-    //   mount(agents);
-    //   helper.click(helper.byTestId(`agent-status-text-${agent.uuid}`));
-    //   expect(helper.byTestId(`agent-build-details-of-${agent.uuid}`)).toBeVisible();
-    //
-    //   helper.click(helper.byTestId(`agent-hostname-of-${agent.uuid}`));
-    //
-    //   expect(helper.byTestId(`agent-build-details-of-${agent.uuid}`)).not.toBeVisible();
-    // });
-
-
-  });
-
   describe("AgentSelection", () => {
     it("should render page with no agent selected", () => {
       mount(agents);
